godweb/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
cloudinary==1.36.0
